requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
Django==3.1.7
django-cors-headers==3.7.0
django-extensions==3.1.3
django-filter==2.4.0
django-jsonfield==1.4.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
filelock==3.0.12
Flask==1.1.2
Flask-Script==2.0.6
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.3
jsonfield==3.1.0
MarkupSafe==1.1.1
mysqlclient==2.0.3
PyJWT==1.7.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
virtualenv==20.4.2
Werkzeug==1.0.1
